requirements.txt
Flask~=3.0.3
python-dotenv~=1.0.1
click~=8.1.7
openai~=1.30.1